migrating `setProps` to be a util that gets imported and then created with `createAppStore`
